+++ packages/desktop/views/dashboard/collectibles/views/CollectiblesGalleryView.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { localize } from '@core/i18n'
-    import { nftSearchTerm, queriedNfts, ownedNfts } from '@core/nfts'
-    import { FontWeight, Illustration, NftGallery, Text, ReceiveButton, SearchInput } from '@ui'
+    import { nftSearchTerm, ownedNfts, queriedNfts } from '@core/nfts/stores'
+    import { FontWeight, Illustration, NftGallery, ReceiveButton, SearchInput, Text } from '@ui'
 </script>
 
 <div class="flex flex-col w-full h-full space-y-4">
