+++ packages/shared/src/lib/core/account/actions/setSelectedAccount.ts
@@ -1,11 +1,10 @@
-import { get } from 'svelte/store'
-import { activeAccounts, updateActiveProfile } from '@core/profile/stores'
-import { resetSendOptionIndex } from '@core/wallet/stores'
-
-import { selectedAccountIndex } from '../stores'
-import { clearFilters } from '@core/utils'
-import { resetNftDownloadQueue } from '@core/nfts'
 import { pollLayer2Tokens } from '@core/layer-2/actions'
+import { resetNftDownloadQueue } from '@core/nfts/stores'
+import { activeAccounts, updateActiveProfile } from '@core/profile/stores'
+import { clearFilters } from '@core/utils'
+import { resetSendOptionIndex } from '@core/wallet/stores'
+import { get } from 'svelte/store'
+import { selectedAccountIndex } from '../stores'
 
 export function setSelectedAccount(index: number): void {
     resetNftDownloadQueue(true)
