+++ packages/shared/src/lib/core/nfts/actions/index.ts
@@ -1,5 +1,6 @@
 export * from './addNftsToDownloadQueue'
 export * from './addOrUpdateNftInAllAccountNfts'
+export * from './buildNftFromNftOutput'
 export * from './downloadNextNftInQueue'
 export * from './getNftByIdFromAllAccountNfts'
 export * from './interruptNftDownloadAfterTimeout'
