+++ packages/shared/src/lib/core/wallet/actions/mintNft.ts
@@ -1,17 +1,18 @@
 import { showNotification } from '@auxiliary/notification'
 import { selectedAccount, updateSelectedAccount } from '@core/account'
+import { ActivityAction } from '@core/activity/enums'
+import { addActivityToAccountActivitiesInAllAccountActivities } from '@core/activity/stores'
+import { NftActivity } from '@core/activity/types'
+import { generateSingleNftActivity } from '@core/activity/utils/generateSingleNftActivity'
+import { preprocessTransaction } from '@core/activity/utils/outputs'
 import { localize } from '@core/i18n'
-import { addOrUpdateNftInAllAccountNfts, buildNftFromNftOutput, IIrc27Metadata } from '@core/nfts'
+import { IIrc27Metadata } from '@core/nfts'
+import { addOrUpdateNftInAllAccountNfts, buildNftFromNftOutput } from '@core/nfts/actions'
 import { Converter } from '@core/utils'
 import { MintNftParams } from '@iota/wallet'
 import { get } from 'svelte/store'
 import { DEFAULT_TRANSACTION_OPTIONS, OUTPUT_TYPE_NFT } from '../constants'
 import { resetMintNftDetails } from '../stores'
-import { preprocessTransaction } from '@core/activity/utils/outputs'
-import { generateSingleNftActivity } from '@core/activity/utils/generateSingleNftActivity'
-import { NftActivity } from '@core/activity/types'
-import { addActivityToAccountActivitiesInAllAccountActivities } from '@core/activity/stores'
-import { ActivityAction } from '@core/activity/enums'
 
 export async function mintNft(metadata: IIrc27Metadata, quantity: number): Promise<void> {
     try {
